Move SmbDirectoryWrapper outside of lucene package into ES package
